fix(pedantic): long long and ;

+++ include/class/parsing_class.hpp
@@ -29,7 +29,7 @@ class Parsing_class {
         int _server_fd;
         int _server_id;
 		std::map<int, std::string> _error_pages;
-		long long _client_max_body_size;
+		long _client_max_body_size;
 		bool _error;
 		std::string _index;
 		std::map<std::string, LocationData> _LocationMap;
@@ -50,7 +50,7 @@ class Parsing_class {
         void setError(bool error);
 		void setMap(const std::string &path, const LocationData &data);
 		void setErrorPage(int error_code, const std::string& page);
-		void setClientMaxBodySize(long long size);
+		void setClientMaxBodySize(long size);
 
         std::string getRoot();
         std::string getName();
@@ -61,7 +61,7 @@ class Parsing_class {
         bool getError();
 		LocationData* getLocation(const std::string& path);
 		std::string getErrorPage(int error_code) ;
-		long long getClientMaxBodySize() const;
+		long getClientMaxBodySize() const;
 
 		// Client management methods
 		void addClient(int client_fd, const ClientData& client);
